# Feature: optimize metrics controller