--- a/controllers/metrics_controller.rb
+++ b/controllers/metrics_controller.rb
@@ -63,6 +63,8 @@ class MetricsController < ApplicationController
   # Display metrics for ontology
   get "/ontologies/:ontology/metrics" do
     check_last_modified_collection(LinkedData::Models::Metric)
+    ont = Ontology.find(params['ontology']).first
+    error 404, "Ontology #{params['ontology']} not found" unless ont
     ontology_metrics = LinkedData::Models::Metric
                       .where(submission: {ontology: [acronym: params['ontology']]})
                       .order_by(submission: {submissionId: :desc})
@@ -72,6 +74,8 @@ class MetricsController < ApplicationController
 
   get "/ontologies/:ontology/submissions/:ontology_submission_id/metrics" do
     check_last_modified_collection(LinkedData::Models::Metric)
+    ont = Ontology.find(params['ontology']).first
+    error 404, "Ontology #{params['ontology']} not found" unless ont
     ontology_submission_metrics = LinkedData::Models::Metric
                       .where(submission: { submissionId: params['ontology_submission_id'].to_i, ontology: [acronym: params['ontology']] })
                       .include(LinkedData::Models::Metric.goo_attrs_to_load(includes_param)).first