### Context
The `/metrics` controller is too slow without the cache. This PR address this issue and optimize the routes.

### Diagnosis
- When we call `/ontologies/STY/metrics` and we check the logs of 4store (triple store) we see a lot of sparql queries are executed
![image](https://github.com/user-attachments/assets/a8c68542-1864-49a0-b404-bb70bcfb2b2a)

- The number of sparql queries executed (from Q2096 to Q2130) = 35 

- Check the full list of queries here: [List sparql queries (metrics)](https://www.notion.so/imadbourouche/Diagnosis-1b22388ee64580bba038fe31a5932cf8)
### Improvement
- With the new code we optimized a lot the number of sparql queries executed 
![image](https://github.com/user-attachments/assets/b377aedb-0db6-4eee-b40a-34674051f88a)

- The number of sparql queries executed (from Q2172 to Q2174) = 3


- Check the full list of queries after this change: [List sparql queries (metrics)](https://www.notion.so/imadbourouche/Diagnosis-1b22388ee64580bba038fe31a5932cf8)

|  criteria  | Before  | After  |
|-----------|-----------|-----------|
| Response time (for /metrics) | 42532 ms  (42 s)  | 14738 ms  (14 s) |
| Number of sparql queries to get metrics of one ontology | 35  queries | 3 queries  |
 
> the response time is tested in lovportal with ~700 ontology

### What's change
- We use now the `LinkedData::Models::Metric` to get the metrics without calling ontologies, submission and doing the logic to get metrics
- routes changed: `/metrics`, `/ontologies/:ontology/metrics`, `/ontologies/:ontology/submissions/:ontology_submission_id/metrics`